+++ R/index_tabular.R
@@ -7,69 +7,70 @@
 #' echotabix}.
 #' 
 #' @param path Path to GWAS summary statistics file.
+#' @param remove_tmp Remove the temporary uncompressed version of the file 
+#' (\emph{.tsv}). 
 #' @param verbose Print messages.
 #' @inheritParams liftover
+#' @inheritParams Rsamtools::bgzip
 #' 
 #' @return Path to tabix-indexed tabular file
 #'
 #' @family tabix
-#' @examples
-#' sumstats_dt <- MungeSumstats::formatted_example()
-#' sumstats_dt <- MungeSumstats:::sort_coords(sumstats_dt = sumstats_dt)
-#' path <- tempfile(fileext = ".tsv")
-#' MungeSumstats::write_sumstats(sumstats_dt = sumstats_dt, save_path = path)
-#' 
-#' indexed_file <- MungeSumstats::index_tabular(path = path)
 #' @export
 #' @importFrom R.utils gunzip isGzipped
+#' @examples
+#' sumstats_dt <- MungeSumstats::formatted_example() 
+#' path <- tempfile(fileext = ".tsv")
+#' MungeSumstats::write_sumstats(sumstats_dt = sumstats_dt, save_path = path)
+#' indexed_file <- MungeSumstats::index_tabular(path = path)
 index_tabular <- function(path,
                           chrom_col = "CHR",
                           start_col = "BP",
                           end_col = start_col,
+                          overwrite = TRUE,
+                          remove_tmp = TRUE,
                           verbose = TRUE) {
     msg <- paste("Converting full summary stats file to",
                  "tabix format for fast querying...")
-    message(msg)
-    #### Read header and make dictionary ####
-    cdict <- column_dictionary(file_path = path)
-    #### Check column exist ####
-    if(!chrom_col %in% names(cdict)) stop("chrom_col not found in file.")
-    if(!start_col %in% names(cdict)) stop("start_col not found in file.")
-    if(!end_col %in% names(cdict)) stop("end_col not found in file.")
+    messager(msg,v=verbose) 
     #### Make sure input file isn't empty ####
+    if(!file.exists(path)){
+        stp <- paste("File does not exist: ",path)
+        stop(stp)
+    }
     if (file.size(path) == 0) {
         msg2 <- paste("Removing empty file:", path)
-        message(msg2)
+        messager(msg2,v=verbose)
         out <- file.remove(path)
     }
     #### Ensure it's not already compressed ####
     if(R.utils::isGzipped(path)){
         path <- R.utils::gunzip(path, overwrite = TRUE)
     }
+    #### Read header and make dictionary ####
+    cdict <- column_dictionary(file_path = path)
+    #### Check column exist ####
+    if(!chrom_col %in% names(cdict)) stop("chrom_col not found in file.")
+    if(!start_col %in% names(cdict)) stop("start_col not found in file.")
+    if(!end_col %in% names(cdict)) stop("end_col not found in file.")
     ### File MUST be bgzipped first
-    message("Ensuring file is bgzipped.")
+    messager("Ensuring file is bgzipped.",v=verbose)
     bgz_file <- Rsamtools::bgzip(file = path,
                                  dest = sprintf("%s.bgz",
                                                 sub("\\.gz$|\\.bgz$", "",
                                                     path)),
-                                 overwrite = TRUE)
+                                 overwrite = overwrite)
     ### Tabix-index file
-    message("Tabix-indexing file.") 
-    # Rsamtools::indexTabix is not user-friendly, 
-    # and is prone to errors without additional wrappers functions
-    #(ie seqminer::tabix.createIndex)
-    # tbx <- Rsamtools::indexTabix(file = bgz_file,
-    #                               seq = chrom_col,
-    #                               start = start_col,
-    #                               comment = "SNP", 
-    #                               end = end_col)
-    seqminer::tabix.createIndex(
-        bgzipFile = bgz_file,
-        sequenceColumn = cdict[[chrom_col]],
-        startColumn = cdict[[start_col]],
-        endColumn = cdict[[end_col]],
-        ## Just use the first columns name
-        metaChar = names(cdict)[1]
-    )
+    messager("Tabix-indexing file.",v=verbose)  
+    tbi_file <- Rsamtools::indexTabix(file = bgz_file, 
+                                      seq = cdict[[chrom_col]],
+                                      start = cdict[[start_col]],
+                                      end = cdict[[end_col]],
+                                      comment = names(cdict)[1]) 
+    if(isTRUE(remove_tmp) && 
+       file.exists(path)){
+        messager("Removing temproary .tsv file.",v=verbose)  
+        out <- file.remove(path)
+    }
     return(bgz_file)
 }