+++ R/validate_parameters.R
@@ -355,10 +355,9 @@ validate_parameters <- function(path,
     }
     
     if(tabix_index && 
-       any(!requireNamespace("Rsamtools", quietly = TRUE),
-           !requireNamespace("seqminer", quietly = TRUE),
+       any(!requireNamespace("Rsamtools", quietly = TRUE), 
            !requireNamespace("MatrixGenerics", quietly = TRUE)) ){
-        pkgs <- c("Rsamtools","seqminer","MatrixGenerics")
+        pkgs <- c("Rsamtools","MatrixGenerics")
         missing_pkgs <- pkgs[!pkgs %in% rownames(utils::installed.packages())]
         tbx_msg <- paste0(
             "The following packages must be installed when tabix_index=TRUE:\n",